refactor(API): getCachePath() moved from middlewares/cache to middlewares/branch
