+++ routes/graph.js
@@ -27,7 +27,7 @@ router.get('/:idBranch/graph',
       .withMessage(createErrMsg.invalidParameter('y')),
   ],
   validateParams,
-  cache.getCachePath,
+  branch.getCachePath,
   cache.getOverviews,
   graph.getGraph,
   pgClient.close,
