+++ routes/patch.js
@@ -61,7 +61,6 @@ router.get('/:idBranch/patches',
       .withMessage(createErrMsg.invalidParameter('idBranch')),
   ],
   validateParams,
-  cache.getCachePath,
   patch.getPatches,
   pgClient.close,
   returnMsg);
@@ -78,7 +77,7 @@ router.post('/:idBranch/patch',
     ...geoJsonAPatcher,
   ],
   validateParams,
-  cache.getCachePath,
+  branch.getCachePath,
   cache.getOverviews,
   patch.postPatch,
   pgClient.close,
@@ -94,7 +93,7 @@ router.put('/:idBranch/patch/undo',
       .withMessage(createErrMsg.invalidParameter('idBranch')),
   ],
   validateParams,
-  cache.getCachePath,
+  branch.getCachePath,
   cache.getOverviews,
   patch.undo,
   pgClient.close,
@@ -110,7 +109,7 @@ router.put('/:idBranch/patch/redo',
       .withMessage(createErrMsg.invalidParameter('idBranch')),
   ],
   validateParams,
-  cache.getCachePath,
+  branch.getCachePath,
   cache.getOverviews,
   patch.redo,
   pgClient.close,
@@ -126,7 +125,7 @@ router.put('/:idBranch/patches/clear',
       .withMessage(createErrMsg.invalidParameter('idBranch')),
   ],
   validateParams,
-  cache.getCachePath,
+  branch.getCachePath,
   cache.getOverviews,
   patch.clear,
   pgClient.close,
